requirements.txt
pyautogui>=0.9
mss>=2.0
neat-python
numpy
pillow
opencv3>=3.1